+++ src/onegov.town/onegov/town/tests/test_cli.py
@@ -1,24 +1,48 @@
+import os
+import yaml
+
 from click.testing import CliRunner
 from onegov.town.cli import cli
 
 
-def test_add_town(postgres_dsn):
+def test_manage_towns(postgres_dsn, temporary_directory):
+
+    cfg = {
+        'applications': [
+            {
+                'path': '/onegov_town/*',
+                'application': 'onegov.town.TownApp',
+                'namespace': 'onegov_town',
+                'configuration': {
+                    'dsn': postgres_dsn
+                }
+            }
+        ]
+    }
+
+    cfg_path = os.path.join(temporary_directory, 'onegov.yml')
+
+    with open(cfg_path, 'w') as f:
+        f.write(yaml.dump(cfg))
 
     runner = CliRunner()
     result = runner.invoke(cli, [
-        '--dsn', postgres_dsn,
-        '--schema', 'test-add-town',
-        'add', 'Govikon'
+        '--config', cfg_path, '/onegov_town/newyork', 'add', 'New York'
     ])
 
     assert result.exit_code == 0
-    assert 'Govikon was added to the test-add-town schema' in result.output
+    assert "New York was created successfully" in result.output
 
     result = runner.invoke(cli, [
-        '--dsn', postgres_dsn,
-        '--schema', 'test-add-town',
-        'add', 'Govikon'
+        '--config', cfg_path, '/onegov_town/newyork', 'add', 'New York'
     ])
 
     assert result.exit_code == 1
-    assert 'The schema test-add-town already contains a town' in result.output
+    assert "/onegov_town/newyork already contains a town" in result.output
+
+    result = runner.invoke(cli, [
+        '--config', cfg_path, '/onegov_town/newyork', 'delete'
+    ], input='y\n')
+
+    assert result.exit_code == 0
+    assert "New York was deleted successfully" in result.output